+++ acvm/src/pwg/directives/mod.rs
@@ -1,10 +1,7 @@
 use std::cmp::Ordering;
 
 use acir::{
-    circuit::{
-        directives::{Directive, LogInfo, QuotientDirective},
-        OpcodeLabel,
-    },
+    circuit::directives::{Directive, QuotientDirective},
     native_types::WitnessMap,
     FieldElement,
 };
@@ -13,7 +10,7 @@ use num_traits::Zero;
 
 use crate::OpcodeResolutionError;
 
-use super::{get_value, insert_value, witness_to_value};
+use super::{get_value, insert_value, ErrorLocation};
 
 mod sorting;
 
@@ -28,11 +25,6 @@ pub(super) fn solve_directives(
     directive: &Directive,
 ) -> Result<(), OpcodeResolutionError> {
     match directive {
-        Directive::Invert { x, result } => {
-            let val = witness_to_value(initial_witness, *x)?;
-            insert_value(result, val.inverse(), initial_witness)?;
-            Ok(())
-        }
         Directive::Quotient(QuotientDirective { a, b, q, r, predicate }) => {
             let val_a = get_value(a, initial_witness)?;
             let val_b = get_value(b, initial_witness)?;
@@ -74,7 +66,7 @@ pub(super) fn solve_directives(
 
             if b.len() < decomposed_integer.len() {
                 return Err(OpcodeResolutionError::UnsatisfiedConstrain {
-                    opcode_label: OpcodeLabel::Unresolved,
+                    opcode_location: ErrorLocation::Unresolved,
                 });
             }
 
@@ -126,55 +118,7 @@ pub(super) fn solve_directives(
             }
             Ok(())
         }
-        Directive::Log(info) => {
-            let witnesses = match info {
-                LogInfo::FinalizedOutput(output_string) => {
-                    println!("{output_string}");
-                    return Ok(());
-                }
-                LogInfo::WitnessOutput(witnesses) => witnesses,
-            };
-
-            if witnesses.len() == 1 {
-                let witness = &witnesses[0];
-                let log_value = witness_to_value(initial_witness, *witness)?;
-                println!("{}", format_field_string(*log_value));
-                return Ok(());
-            }
-
-            // If multiple witnesses are to be fetched for a log directive,
-            // it assumed that an array is meant to be printed to standard output
-            //
-            // Collect all field element values corresponding to the given witness indices
-            // and convert them to hex strings.
-            let mut elements_as_hex = Vec::with_capacity(witnesses.len());
-            for witness in witnesses {
-                let element = witness_to_value(initial_witness, *witness)?;
-                elements_as_hex.push(format_field_string(*element));
-            }
-
-            // Join all of the hex strings using a comma
-            let comma_separated_elements = elements_as_hex.join(", ");
-
-            let output_witnesses_string = "[".to_owned() + &comma_separated_elements + "]";
-
-            println!("{output_witnesses_string}");
-
-            Ok(())
-        }
     }
-}
-
-/// This trims any leading zeroes.
-/// A singular '0' will be prepended as well if the trimmed string has an odd length.
-/// A hex string's length needs to be even to decode into bytes, as two digits correspond to
-/// one byte.
-fn format_field_string(field: FieldElement) -> String {
-    let mut trimmed_field = field.to_hex().trim_start_matches('0').to_owned();
-    if trimmed_field.len() % 2 != 0 {
-        trimmed_field = "0".to_owned() + &trimmed_field
-    }
-    "0x".to_owned() + &trimmed_field
 }
 
 #[cfg(test)]