+++ blackbox_solver/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## [0.24.1](https://github.com/noir-lang/acvm/compare/acvm_blackbox_solver-v0.24.0...acvm_blackbox_solver-v0.24.1) (2023-09-03)
+
+
+### Bug Fixes
+
+* Add WASI 20 `_initialize` call to `acvm_backend.wasm` binary ([#518](https://github.com/noir-lang/acvm/issues/518)) ([ec6ab0c](https://github.com/noir-lang/acvm/commit/ec6ab0c6fb2753209abe1e03a449873e255ffd76))
+
+## [0.24.0](https://github.com/noir-lang/acvm/compare/acvm_blackbox_solver-v0.23.0...acvm_blackbox_solver-v0.24.0) (2023-08-31)
+
+
+### Miscellaneous Chores
+
+* **acvm_blackbox_solver:** Synchronize acvm versions
+
+## [0.23.0](https://github.com/noir-lang/acvm/compare/acvm_blackbox_solver-v0.22.0...acvm_blackbox_solver-v0.23.0) (2023-08-30)
+
+
+### ⚠ BREAKING CHANGES
+
+* **acvm:** Remove `BlackBoxFunctionSolver` from `Backend` trait ([#494](https://github.com/noir-lang/acvm/issues/494))
+
+### Features
+
+* Expose a `BlackBoxFunctionSolver` containing a barretenberg wasm from `blackbox_solver` ([#494](https://github.com/noir-lang/acvm/issues/494)) ([a1d4b71](https://github.com/noir-lang/acvm/commit/a1d4b71256dfbf1e883e770dd9c45479235aa860))
+
+
+### Miscellaneous Chores
+
+* **acvm:** Remove `BlackBoxFunctionSolver` from `Backend` trait ([#494](https://github.com/noir-lang/acvm/issues/494)) ([a1d4b71](https://github.com/noir-lang/acvm/commit/a1d4b71256dfbf1e883e770dd9c45479235aa860))
+
+## [0.22.0](https://github.com/noir-lang/acvm/compare/acvm_blackbox_solver-v0.21.0...acvm_blackbox_solver-v0.22.0) (2023-08-18)
+
+
+### Miscellaneous Chores
+
+* **acvm_blackbox_solver:** Synchronize acvm versions
+
 ## [0.21.0](https://github.com/noir-lang/acvm/compare/acvm_blackbox_solver-v0.20.1...acvm_blackbox_solver-v0.21.0) (2023-07-26)